# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/app/snapshot/create": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "app快照"
                ],
                "summary": "app创建存储快照",
                "parameters": [
                    {
                        "description": "app创建存储快照",
                        "name": "\"\"",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.SnapshotCreateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Result"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Result"
                        }
                    }
                }
            }
        },
        "/app/snapshot/delete": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "app快照"
                ],
                "summary": "删除快照",
                "parameters": [
                    {
                        "description": "删除快照",
                        "name": "\"\"",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.SnapshotDeleteReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Result"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Result"
                        }
                    }
                }
            }
        },
        "/app/snapshot/list": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "app快照"
                ],
                "summary": "获取快照列表",
                "parameters": [
                    {
                        "description": "获取快照列表",
                        "name": "\"\"",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.SnapshotListReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.SnapshotListRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Result"
                        }
                    }
                }
            }
        },
...
        "/helm/release/Update": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "helm release"
                ],
                "summary": "helm release更新",
                "parameters": [
                    {
                        "description": "helm release更新",
                        "name": "\"\"",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.HelmReleaseUpdateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Result"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Result"
                        }
                    }
                }
            }
        },
        "/helm/release/create": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "helm release"
                ],
                "summary": "helm release创建",
                "parameters": [
                    {
                        "description": "helm release创建",
                        "name": "\"\"",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.HelmReleaseCreateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Result"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Result"
                        }
                    }
                }
            }
        },
        "/helm/release/delete": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "helm release"
                ],
                "summary": "helm release删除",
                "parameters": [
                    {
                        "description": "helm release删除",
                        "name": "\"\"",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.HelmReleaseReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Result"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Result"
                        }
                    }
                }
            }
        },
        "/helm/release/list": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "helm release"
                ],
                "summary": "helm release列表信息",
                "parameters": [
                    {
                        "description": "helm release列表信息",
                        "name": "\"\"",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.HelmReleaseListReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.HelmReleaseListRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Result"
                        }
                    }
                }
            }
        },
        "/helm/release/values": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "helm release"
                ],
                "summary": "获取helm release的values",
                "parameters": [
                    {
                        "description": "获取helm release的values",
                        "name": "\"\"",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.HelmReleaseReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Result"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Result"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.HelmApp": {
            "type": "object",
            "properties": {
                "chart_name": {
                    "type": "string"
                },
                "create_time": {
                    "type": "string"
                },
                "end_point": {
                    "type": "object"
                },
                "id": {
                    "type": "integer"
                },
                "modify_time": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "repo_name": {
                    "type": "string"
                },
                "repo_url": {
                    "type": "string"
                },
                "status": {
                    "description": "值为0的时候显示创建中(灰色转圈图标),值为-1表示创建失败(红色失败图标)，值为1表示创建成功(绿色成功图标)",
                    "type": "integer"
                },
                "user_name": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "model.HelmReleaseCreateReq": {
            "type": "object",
            "required": [
                "chart_name",
                "name",
                "repo_name",
                "repo_url",
                "value",
                "version"
            ],
            "properties": {
                "chart_name": {
                    "description": "对应name",
                    "type": "string"
                },
                "name": {
                    "description": "名称,需要用户输入,后端做了唯一性校验",
                    "type": "string"
                },
                "repo_name": {
                    "description": "对应repository.name",
                    "type": "string"
                },
                "repo_url": {
                    "description": "对应repository.url",
                    "type": "string"
                },
                "value": {
                    "description": "用户在json编辑器输入的内容",
                    "type": "string"
                },
                "version": {
                    "description": "对应version",
                    "type": "string"
                }
            }
        },
        "model.HelmReleaseList": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.HelmApp"
                    }
                }
            }
        },
        "model.HelmReleaseListReq": {
            "type": "object",
            "required": [
                "current",
                "pageSize"
            ],
            "properties": {
                "current": {
                    "description": "当前第几页",
                    "type": "integer"
                },
                "pageSize": {
                    "description": "每页多少条",
                    "type": "integer"
                },
                "search": {
                    "description": "模糊搜索字段，可根据名称来搜索",
                    "type": "string"
                },
                "user_name": {
                    "description": "用户名",
                    "type": "string"
                }
            }
        },
        "model.HelmReleaseListRes": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/model.HelmReleaseList"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "model.HelmReleaseReq": {
            "type": "object",
            "properties": {
                "name": {
                    "description": "release名称",
                    "type": "string"
                }
            }
        },
        "model.HelmReleaseUpdateReq": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "chart_name": {
                    "type": "string"
                },
                "name": {
                    "description": "用户输入的，需要全局唯一",
                    "type": "string"
                },
                "repo_name": {
                    "type": "string"
                },
                "repo_url": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
...
        "model.Result": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "model.SnapshotCreateReq": {
            "type": "object",
            "required": [
                "app_name",
                "show_name"
            ],
            "properties": {
                "app_name": {
                    "description": "要创建快照的app名称，注意不是show_name",
                    "type": "string"
                },
                "show_name": {
                    "description": "名称",
                    "type": "string"
                }
            }
        },
        "model.SnapshotDeleteReq": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "description": "快照id",
                    "type": "integer"
                }
            }
        },
        "model.SnapshotInfo": {
            "type": "object",
            "properties": {
                "app_name": {
                    "type": "string"
                },
                "app_show_name": {
                    "type": "string"
                },
                "create_time": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "modify_time": {
                    "type": "string"
                },
                "name": {
                    "description": "快照名称,对应k8s snapshot的资源名",
                    "type": "string"
                },
                "show_name": {
                    "description": "给用户展示的名称",
                    "type": "string"
                },
                "user_name": {
                    "type": "string"
                }
            }
        },
        "model.SnapshotListBase": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.SnapshotInfo"
                    }
                }
            }
        },
        "model.SnapshotListReq": {
            "type": "object",
            "required": [
                "current",
                "pageSize"
            ],
            "properties": {
                "app_name": {
                    "description": "app名称，可不传",
                    "type": "string"
                },
                "current": {
                    "description": "当前第几页",
                    "type": "integer"
                },
                "pageSize": {
                    "description": "每页多少条",
                    "type": "integer"
                },
                "search": {
                    "description": "模糊搜索字段，可不传",
                    "type": "string"
                },
                "user_name": {
                    "description": "用户名,可不传",
                    "type": "string"
                }
            }
        },
        "model.SnapshotListRes": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/model.SnapshotListBase"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
    }
}